+++ layers/CMSSchema/packages/custompost-mutations-wp/src/TypeAPIs/CustomPostTypeMutationAPI.php
@@ -41,6 +41,16 @@ class CustomPostTypeMutationAPI extends AbstractBasicService implements CustomPo
             $query['ID'] = $query['id'];
             unset($query['id']);
         }
+        if (isset($query['parent-id'])) {
+            $query['post_parent'] = $query['parent-id'];
+            unset($query['parent-id']);
+        }
+        // Passing `0` as parent means "remove the parent", so it's already handled above
+        // elseif (array_key_exists('parent-id', $query)) {
+        //     // If passing `null` then remove the parent
+        //     $query['post_parent'] = 0;
+        //     unset($query['parent-id']);
+        // }
         if (isset($query['content'])) {
             $query['post_content'] = $query['content'];
             unset($query['content']);